state: move Validator methods into validators.go
